further changes in order to realize the plugin documentation

+++ docs/pluginDoc/create_plugin_doc.py
@@ -16,11 +16,11 @@ import __main__
 import shutil
 import glob
 
-def process_docstring(app, what, name, obj, options, lines):
-    pass
+#def process_docstring(app, what, name, obj, options, lines):
+#    pass
     
-def process_signature(app, what, name, obj, options, signature, return_annotation):
-    pass
+#def process_signature(app, what, name, obj, options, signature, return_annotation):
+#    pass
     
 
 buildernames = ["qthelp"] #["qthelp", "htmlhelp", "latex", "html"]
@@ -45,7 +45,7 @@ def createPluginDoc(confFile, buildernames):
     all_files = True
     filenames = False
     confoverrides = {}
-    freshenv = False
+    freshenv = True #fresh environment variable, else False
 
     basedir = itom.getCurrentPath()
     srcdir = cfgDict["pluginDocSourceDir"] #from pluginConfiguration
@@ -55,48 +55,52 @@ def createPluginDoc(confFile, buildernames):
         outdir = pathConv(os.path.join(cfgDict["pluginDocBuildDir"],buildername))
         doctreedir = os.path.join(cfgDict["pluginDocBuildDir"],"doctrees")
         
-        helpDict = itom.pluginHelp(cfgDict["pluginDocTarget"],True)
-        
-        confoverrides = {"project": helpDict["name"],
-            "copyright": helpDict["author"],
-            "project": "itom plugin '" + helpDict["name"] + "'",
-            "version": helpDict["version"],
-            "release": "",
-            "html_title": helpDict["name"] + " (" + helpDict["type"] + ")",
-            "html_short_title": helpDict["name"],
-            "master_doc": cfgDict["pluginDocMainDocument"] }
-        
-        app = Sphinx(srcdir, confdir, outdir, doctreedir, buildername,
-                     confoverrides, sys.stdout, sys.stderr, freshenv)
-        
-        app.connect('autodoc-process-docstring',process_docstring)
-        app.connect('autodoc-process-signature',process_signature)
-        
-        try:
-            os.mkdir(outdir)
-        except:
-            pass
-
-        nocolor()
-
-        if not app.builder:
-            raise RuntimeError
-
-        if all_files:
-            app.builder.build_all()
-        elif filenames:
-            app.builder.build_specific(filenames)
-        else:
-            app.builder.build_update()
-        
-        if (buildername == "qthelp"):
-            #copy important files from qthelp subfolder to pluginDocInstallDir
-            pluginDocInstallDir = pathConv(cfgDict["pluginDocInstallDir"])
+        if (itom.pluginLoaded(cfgDict["pluginDocTarget"])):
+            helpDict = itom.pluginHelp(cfgDict["pluginDocTarget"],True)
             
-            if (os.path.exists(pluginDocInstallDir)):
-                shutil.rmtree(pluginDocInstallDir)
-                    
-            shutil.copytree(outdir, pluginDocInstallDir, ignore=shutil.ignore_patterns("*.js","search.html",".buildinfo"))
+            confoverrides = {"project": helpDict["name"],
+                "copyright": helpDict["author"],
+                "project": "itom plugin '" + helpDict["name"] + "'",
+                "version": helpDict["version"],
+                "release": "",
+                "html_title": helpDict["name"] + " (" + helpDict["type"] + ")",
+                "html_short_title": helpDict["name"],
+                "master_doc": cfgDict["pluginDocMainDocument"] }
+        
+            app = Sphinx(srcdir, confdir, outdir, doctreedir, buildername,
+                         confoverrides, sys.stdout, sys.stderr, freshenv)
+            
+            #app.connect('autodoc-process-docstring',process_docstring)
+            #app.connect('autodoc-process-signature',process_signature)
+            
+            try:
+                os.mkdir(outdir)
+            except:
+                pass
+
+            nocolor()
+
+            if not app.builder:
+                raise RuntimeError
+
+            if all_files:
+                app.builder.build_all()
+            elif filenames:
+                app.builder.build_specific(filenames)
+            else:
+                app.builder.build_update()
+            
+            if (buildername == "qthelp"):
+                #copy important files from qthelp subfolder to pluginDocInstallDir
+                pluginDocInstallDir = pathConv(cfgDict["pluginDocInstallDir"])
+                
+                if (os.path.exists(pluginDocInstallDir)):
+                    shutil.rmtree(pluginDocInstallDir)
+                        
+                shutil.copytree(outdir, pluginDocInstallDir, ignore=shutil.ignore_patterns("*.js","search.html",".buildinfo"))
+        
+        else:
+            print("Plugin documentation for", helpDict["name"], "could not be build since not available on this computer")
             
 
 if (__name__ == "__main__"):
